Add remaining parts of the README.md
The README was incomplete at this stage.

The incomplete README makes it hard for new users to navigate the
usage of our product.

Adding the 2nd part of the README helps complete the README file.

Splitting into 2 different parts allows 2 developers to work on this
and allows for more efficient workflow.