Add view commands for vendors in UG-features
Features in the UG were outdated.

This will cause users to be confused about the usage of WedLog.

Inputting the updated features specifically viewing all vendors
and view a specific vendor will help users navigate WedLog better.
Other parts of the UG were left untouched.

This is so when other edits are made in the future, us developers
can reference the other sections that were previously written
and not waste past efforts.