Add Appendix A: Miscellaneous error message to User Guide
The user guide does not specify expected behavior when WedLog is faced
with erroneous inputs that do not fall under a specific command.

When users are faced with these behaviours, they may not understand
the reason why.

Let's add a section that details the expected output for different
types of input errors that are not tied to a specific command.