DietaryRequirements: Fix bugs
The constructor does not trim the requirement string because the parser already does so. There are no issues with the current workflow.

However, this is not intuitive. Passing in strings with only spaces into the constructor results in dietary requirements of status present that contain whitespace values. This may make the code  and testing more difficult to maintain.

Let's,
* Add another trim to make the constructor behaviour more predictable.
* Fix equals() null cases