Wrap `Phone`, `Email`, `Address` in `Optional`
Phone, Email, and Address are optional fields in Person, represented as null if the Person does not have that field.

Let's wrap the fields in Optional, e.g. from Phone to Optional<Phone>. If the Person does not have that field, it is stored as Optional#empty().

This makes it clearer which fields in Person are optional, and enforces that the absence of the field is handled in relevant methods.