Wrap `TableNumber` in `Optional`
TableNumber is an optional field in Person, represented as null if the Person does not have that field.

Let's wrap the field in Optional. If the Person does not have a table number, it is stored as Optional#empty().

This makes it clearer which fields in Person are optional, and enforces that the absence of the field is handled in relevant methods.