VendorEditCommand: Make fields deletable
The application only allows edits which update vendor field values, and does not support deleting vendor fields which are optional.

Let's,
* Update VendorEditCommand to support deleting optional fields
* Update VendorEditCommandParser to accept the relevant arguments

This will allow the user to key in `vendor edit 1 p/`, which will remove the phone field on the first vendor.

Note that new relevant test cases have not been added yet.