Model: Update predicates to be more general
Some of the predicates such as GuestNamePredicate apply to any Person, hence it should not be implementing Predicate<Guest> but Predicate<Person> instead. This results in duplicated code, such as between GuestNamePredicate and VendorNamePredicate.

Unifying the duplicated code improves the code quality, and reduces the number of tests we have to write.

Let's refactor common person predicate classes into their own classes that implement Predicate<Person>, and remove the former common predicate classes.