Command: Refactor FilterCommand fields
Equality between FilterCommand is problematic as the predicates they are derived from are chained.

Let's modify predicates to be a list within FilterCommand, and delay the creation of a chained predicate from the Parser class to within the FilterCommand class itself.

This allows us to use equality of FilterCommand objects within our tests.