Name: Make name equality case insensitive
Name is case sensitive, which means that two names like `John` and `john` are considered different names.

This is inconsistent with other sections of the code base, such as the `GuestFilterCommand` which is case-insensitive. It is also unintuitive.

Let's update the equality of name to be case insensitive.

This will also update the functionality of Person#isSamePerson, preventing users from adding two users with full names that only differ in cases.